INC-878 Switch feature
