+++ src/test/resources/application-test-nomis.yml
@@ -1,0 +1,2 @@
+feature:
+  incentives-data-source-of-truth: false
