+++ src/test/resources/application-test.yml
@@ -47,3 +47,6 @@ hmpps.sqs:
   topics:
     domainevents:
       arn: arn:aws:sns:eu-west-2:000000000000:${random.uuid}
+
+feature:
+  incentives-data-source-of-truth: true
